Added more comments and added MirroredStrategy dispatching in edge prediction pipeline

+++ tests/test_evaluate_embedding_for_edge_prediction.py
@@ -1,12 +1,10 @@
 """Unit test class for GraphTransformer objects."""
 from unittest import TestCase
 from ensmallen import Graph  # pylint: disable=no-name-in-module
-from embiggen import GraphVisualization, GloVe
 from embiggen.pipelines import evaluate_embedding_for_edge_prediction
-from plot_keras_history import plot_history
-import matplotlib.pyplot as plt
 from ensmallen.datasets.networkrepository import Karate
 import shutil
+import os
 
 
 class TestEvaluateEmbeddingForEdgePrediction(TestCase):
@@ -15,12 +13,15 @@ class TestEvaluateEmbeddingForEdgePrediction(TestCase):
     def setUp(self):
         """Setup objects for running tests on GraphTransformer objects class."""
         self._graph = Karate()
-        self._subgraph = self._graph.get_random_subgraph(self._graph.get_nodes_number() // 2)
+        self._subgraph = self._graph.get_random_subgraph(
+            self._graph.get_nodes_number() - 2
+        ).drop_singleton_nodes()
         self._number_of_holdouts = 2
 
     def test_evaluate_embedding_for_edge_prediction(self):
         """Test graph visualization."""
-        shutil.rmtree("node_embeddings")
+        if os.path.exists("node_embeddings"):
+            shutil.rmtree("node_embeddings")
         holdouts, histories = evaluate_embedding_for_edge_prediction(
             embedding_method="GloVe",
             graph=self._graph,
@@ -38,7 +39,8 @@ class TestEvaluateEmbeddingForEdgePrediction(TestCase):
 
     def test_evaluate_embedding_for_edge_prediction_in_subgraph(self):
         """Test graph visualization."""
-        shutil.rmtree("node_embeddings")
+        if os.path.exists("node_embeddings"):
+            shutil.rmtree("node_embeddings")
         holdouts, histories = evaluate_embedding_for_edge_prediction(
             embedding_method="GloVe",
             graph=self._graph,
